refactor: Split bridge module
